Feat: quiz sessions (#2)
* feat: Add socket io clients and docker for local env

New files for running the project in a local environment where added.

* feat: integrate redis streams

For the quizzes to interact in real team with multiple servers and to persist information the redis streams plug in was added to the socket io server.

* feat: create redis adapter

A personal adapter folder for the backend was created to maintain code that make connections outside the server.

* feat: implement stream reader for testing

The class that reads a stream from Redis was implemented to get and filter a stream message from socket io.

* feat: refactor redis integration

Previously the socket ios stream adapter was used to integrate Redis. A new logic was introduced to make it easier to get and set messages into a stream for a quiz session.

* feat: add test socket io clients

Implemented quiz joining and position setting in socketIOAdmin and handled quiz position reception in socketIOClient

* feat: Add README and clean logs

* fix: remove console log

* fix: add redis url env variable

* refactor: redis env variables

The redis DB connections parameters were separated into 4 env variables: host, port, user and password. Also, dotenv package was added and the dependency to use node 21.

---------

Co-authored-by: Joaquin Popoca <joaquin.popoca@wizeline.com>